pyproject.toml
[tool.poetry]
name = "django-bitza"
version = "0.1.0"
description = ""
authors = ["Сергей Сушков <sushkov_sv@ursip.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
whitenoise = "^6.6.0"
environs = {extras = ["django"], version = "^11.0.0"}
psycopg2 = "^2.9.9"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
pillow = "^10.2.0"
uvicorn = "^0.28.1"
gunicorn = "^21.2.0"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
grpcio = "^1.69.0"
google = "^3.0.0"
protobuf = "^5.29.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
